+++ changelog.md
@@ -1,7 +1,11 @@
 # Changelog
 
+## unreleased
+
 - `rio.FilePickerArea` now reports clicks on uploaded files, so they can e.g. be
   used to download them again
+- Add `on_message` to `rio.Webview`, allowing communication between JavaScript
+  and Python
 
 ## 0.12.1